Add AIP-1: AIP Purpose and Guidelines (#1)
AIP stands for Atomicals Improvement Proposal. An AIP is a design
document providing information to the Atomicals community, or describing
a new feature for Atomicals or its processes or environment.

We intend AIPs to be the primary mechanisms for proposing new features,
for collecting community technical input on an issue, and for
documenting the design decisions that have gone into Atomicals. Because
the AIPs are maintained as text files in a versioned repository, their
revision history is the historical record of the feature proposal.

For Atomicals implementers, AIPs are a convenient way to track the
progress of their implementation. Ideally each implementation maintainer
would list the AIPs that they have implemented. This will give end users
a convenient way to know the current status of a given implementation or
library.

---------

Signed-off-by: Wizz Wallet <153743376+wizz-wallet@users.noreply.github.com>
Co-authored-by: Wizz Wallet <153743376+wizz-wallet@users.noreply.github.com>